Team service authorization

+++ Server/Services/Team/TeamService.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Server.DAOs;
 using Server.Models.Dapper;
 using Server.Services.Authentication;
+using Server.Services.Authorization;
 using Server.Utils;
 
 namespace Server.Services.Team
@@ -14,43 +15,61 @@ namespace Server.Services.Team
 
         private readonly TeamDAO _teamDao;
         private readonly TeamMemberDAO _teamMemberDao;
-        private readonly IUserProvider _userProvider;
+        private readonly IUserAuthorization _userAuthorization;
+        private readonly int _userId;
 
-        public TeamService(IUserProvider userProvider, TeamDAO teamDao, TeamMemberDAO teamMemberDao)
+        public TeamService(IUserProvider userProvider, TeamDAO teamDao, TeamMemberDAO teamMemberDao, IUserAuthorization userAuthorization)
         {
-            _userProvider = userProvider;
             _teamDao = teamDao;
             _teamMemberDao = teamMemberDao;
+            _userAuthorization = userAuthorization;
+            _userId = userProvider.GetUserId();
         }
 
         public async Task<List<TeamBase>> GetTeams()
         {
-            int userId = _userProvider.GetUserId();
-            return await _teamDao.GetTeams(userId);
+            return await _teamDao.GetTeams(_userId);
         }
 
         public async Task<bool> ChangeName(int teamId, string newName)
         {
-            var result = await _teamDao.ChangeName(teamId, newName);
-            return result;
+            var hasPermissions = await _userAuthorization.Authorize(_userId, teamId, Action.RenameTeam);
+
+            if (hasPermissions)
+            {
+                return await _teamDao.ChangeName(teamId, newName);
+            }
+
+            throw new UnauthorizedAccessException();
         }
 
         public async Task<Optional<string>> GenerateJoinCode(int teamId)
         {
-            var code = await _teamDao.GenerateJoinCode(teamId);
-            return code;
+            var hasPermissions = await _userAuthorization.Authorize(_userId, teamId, Action.GenerateJoinCode);
+
+            if (hasPermissions)
+            {
+                return await _teamDao.GenerateJoinCode(teamId);
+            }
+
+            throw new UnauthorizedAccessException();
         }
 
         public async Task<bool> RemoveJoinCode(int teamId)
         {
-            var result = await _teamDao.RemoveJoinCode(teamId);
-            return result;
+            var hasPermissions = await _userAuthorization.Authorize(_userId, teamId, Action.RemoveJoinCode);
+
+            if (hasPermissions)
+            {
+                return await _teamDao.RemoveJoinCode(teamId);
+            }
+
+            throw new UnauthorizedAccessException();
         }
 
         public async Task<Optional<TeamBase>> CreateTeam(string name)
         {
-            int userId = _userProvider.GetUserId();
-            var teamId = await _teamDao.CreateTeam(name, userId);
+            var teamId = await _teamDao.CreateTeam(name, _userId);
 
             if (teamId.IsEmpty)
             {
@@ -66,33 +85,55 @@ namespace Server.Services.Team
 
         public async Task<List<User>> GetMembers(int teamId)
         {
-            var members = await _teamDao.GetMembers(teamId);
-            return members;
+            var hasPermissions = await _userAuthorization.Authorize(_userId, teamId, Action.MemberAccess);
+
+            if (hasPermissions)
+            {
+                return await _teamDao.GetMembers(teamId);
+            }
+
+            throw new UnauthorizedAccessException();
         }
 
         public async Task<bool> AddMember(int teamId, string email)
         {
-            var response = await _teamMemberDao.AddMember(teamId, email);
-            return response;
+            var hasPermissions = await _userAuthorization.Authorize(_userId, teamId, Action.AddMember);
+
+            if (hasPermissions)
+            {
+                return await _teamMemberDao.AddMember(teamId, email);
+            }
+
+            throw new UnauthorizedAccessException();
         }
 
         public async Task<bool> RemoveMember(int teamId, int userId)
         {
-            var response = await _teamMemberDao.RemoveMember(teamId, userId);
-            return response;
+            var hasPermissions = await _userAuthorization.Authorize(_userId, teamId, Action.RemoveMember);
+
+            if (hasPermissions)
+            {
+                return await _teamMemberDao.RemoveMember(teamId, userId);
+            }
+
+            throw new UnauthorizedAccessException();
         }
 
         public async Task<Optional<int>> JoinWithCode(string code)
         {
-            int userId = _userProvider.GetUserId();
-            var response = await _teamMemberDao.JoinWithCode(userId, code);
-            return response;
+            return await _teamMemberDao.JoinWithCode(_userId, code);
         }
 
         public async Task<bool> ChangeUserRole(int teamId, int userId, int role)
         {
-            var response = await _teamMemberDao.ChangeUserRole(teamId, userId, role);
-            return response;
+            var hasPermissions = await _userAuthorization.Authorize(_userId, teamId, Action.ChangeRole);
+
+            if (hasPermissions)
+            {
+                return await _teamMemberDao.ChangeUserRole(teamId, userId, role);
+            }
+
+            throw new UnauthorizedAccessException();
         }
     }
 }
